ci(data-checks): remove the vacuous warehouse stub instead of shipping a TODO
The dbt-and-warehouse-gates job never did anything real: its dbt step was an
'echo' TODO and the audit ran against an empty throwaway ClickHouse, where every
resolution/dedup/presence check passes by construction. A gate that can't fail is
worse than no gate. If it's useless, it shouldn't exist — so remove it.

- data-checks.yml: drop the dbt-and-warehouse-gates job, the schedule trigger, and
  the TODO stub entirely. Keep connector-silver-coverage (a real, warehouse-free
  gate) and document where the warehouse audit actually belongs.
- data_presence_audit.py: drop the --skip-if-unreachable flag + reachable-but-empty
  skip (they only existed to make the vacuous gate 'honest'). The tool runs against
  a real populated warehouse (e2e rig / deployed), so an unreachable warehouse is a
  hard failure again. Keep the https-when-credentialed hardening. Docstring now
  points resolution/dedup at the e2e rig (which dbt-builds), not a standalone job.
- render_contract.py: drop redundant int() casts (math.floor/ceil already return int).

The test_live_render_e2e.py 'use response.items' review note is a false positive:
BatchQueryResult is #[serde(tag="status")] with #[serde(flatten)] response, so
status/items are top-level (verified in domain/query.rs) — the test is correct.

+++ tests/ui_render_contract/render_contract.py
@@ -33,7 +33,7 @@ COMING_SOON = "ComingSoon"
 
 def round_half_up(value: float) -> int:
     """Round to nearest integer, halves away from zero (4.5→5, -4.5→-5, 4.4→4)."""
-    return int(math.floor(value + 0.5)) if value >= 0 else int(math.ceil(value - 0.5))
+    return math.floor(value + 0.5) if value >= 0 else math.ceil(value - 0.5)
 
 
 def display_value(
